sendfromaddress.sh
script that can be used to create unsigned raw transaction to
withdraw from address, based only on explorer API. it can be useful
in cases when you want to withdraw all mined / existing funds from
your NN address, but already did a wallet reset and doesn't have
utxos from past in the wallet. of course, after create tx needs to
be signed and broadcast.

also this example script contains a round fuction written in jq, which
can be useful in other scripts.