Calls split_nn_sapling.sh if low utxo count on NN
I used FSM branch of komodo on the distant KMD node. I tested this script a little bit to split utxo for the 3P NN setup, it worked but I think that we should avoid to rely on only one explorer : if the explorer returns 0 utxo for the NN, we will split loads of utxo, it is not useful. A possible enhancement is to use several explorers (I'd like to avoid to importprivkey of the NN on the distant node).